Split Vercel helpers to match the SDK namespaces
